Solution to Input Two Integers and Print Their Sum

+++ Input Two Integers and Print Their Sum@@6788d5012db25efec946f93d/code.c
@@ -3,7 +3,7 @@ int main(){
 int a,b ;
 scanf("%d",&a);
 scanf("%d",&b);
-float s = a+b;
+int s = a+b;
 printf("Sum: %d",s);
 
     return 0;
